pyproject.toml
[project]
name = "mcp-practice-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.14.1",
    "openai-agents>=0.3.1",
]
